separate redis from app.js
