+++ app.js
@@ -12,41 +12,7 @@ const pug = require('pug');
 const compression = require('compression');
 const express = require('express');
 const cookieParser = require('cookie-parser');
-const r = require('redis');
-
-const redis = (() => {
-  if (!config.redis_enabled) {
-    // Stub Redis if disabled
-    return {
-      get: (_, callback) => callback(null, null),
-      setex: (_, _1, _2, callback) => callback(null),
-      on: () => {},
-    };
-  }
-
-  const redisOptions = {
-    host: '127.0.0.1',
-    port: 6379,
-  };
-
-  if (config.redis_db) {
-    redisOptions.db = config.redis_db;
-  }
-
-  if (config.redis_host) {
-    redisOptions.host = config.redis_host;
-  }
-
-  if (config.redis_port && config.redis_port > 0) {
-    redisOptions.port = config.redis_port;
-  }
-
-  if (config.redis_password) {
-    redisOptions.password = config.redis_password;
-  }
-
-  return r.createClient(redisOptions);
-})();
+const { redis } = require('./inc/redis');
 
 const nodeFetch = require('node-fetch');
 const fetch = config.http_proxy
@@ -161,12 +127,6 @@ if (config.redirect_http_to_https) {
     else res.redirect(`https://${req.headers.host}${req.url}`);
   });
 }
-
-redis.on('error', (error) => {
-  if (error) {
-    console.error(`Redis error: ${error}`);
-  }
-});
 
 const cacheControl = require('./cacheControl.js');
 cacheControl.removeCacheFiles();
